Use Msa loader + improve look and feel + add some user perms

+++ index.js
@@ -20,13 +20,32 @@ class MsaIdeasModule extends Msa.Module {
 		this.vote = new MsaVoteModule(this.dbKey)
 	}
 
-	getFullDbKey(key){
+	getDbKey(key){
 		return this.dbKey + '-' + key
+	}
+
+	getUserKey(req){
+		const user = req.session ? req.session.user : null
+		return user ? user.name : req.connection.remoteAddress
+	}
+
+	canUserEditIdea(idea, userKey){
+		return (idea.createdBy == userKey)
+	}
+
+	canUserRemoveIdea(idea, userKey){
+		return (idea.createdBy == userKey)
+	}
+
+	completeIdea(idea, req){
+		idea.canEdit = this.canUserEditIdea(idea, this.getUserKey(req))
+		idea.canRemove = this.canUserRemoveIdea(idea, this.getUserKey(req))
 	}
 
 	initApp(){
 		const app = this.app
 
+		// get page
 		app.get("/:key", (req, res, next) => {
 			const key = req.params.key
 			if(key.indexOf('-') >= 0 || key[0] === '_')
@@ -42,21 +61,24 @@ class MsaIdeasModule extends Msa.Module {
 
 		// list ideas
 		app.get("/_list/:key", async (req, res, next) => {
-			const key = this.getFullDbKey(req.params.key)
-			const ideas = await this.db.findAll({ where:{ key }})
+			const key = this.getDbKey(req.params.key)
+			const ideas = (await this.db.findAll({ where:{ key }}))
+				.map(idea => idea.dataValues)
+			ideas.forEach(idea => this.completeIdea(idea, req))
 			const votes = await this.vote.getVoteCounts(key)
 			res.json({ ideas, votes })
-	})
+		})
 
 		// post new idea
 		app.post("/_idea/:key", async (req, res, next) => {
 			try {
-				const key = this.getFullDbKey(req.params.key),
+				const key = this.getDbKey(req.params.key),
 					text = req.body.text,
 					parent = req.body.parent
 				const maxNum = await this.db.max('num', { where:{ key }})
 				const num = Number.isNaN(maxNum) ? 0 : (maxNum+1)
-				await this.db.create({ key, num, text, parent })
+				const createdBy = this.getUserKey(req)
+				await this.db.create({ key, num, text, parent, createdBy })
 				res.sendStatus(200)
 			} catch(err) { next(err) }
 		})
@@ -64,8 +86,12 @@ class MsaIdeasModule extends Msa.Module {
 		// delete idea
 		app.delete("/_idea/:key/:num", async (req, res, next) => {
 			try {
-				const key = this.getFullDbKey(req.params.key),
+				const key = this.getDbKey(req.params.key),
 					num = req.params.num
+				const userKey = this.getUserKey(req)
+				const idea = await this.db.find({ where:{ key, num }})
+				if(!this.canUserRemoveIdea(idea, userKey))
+					return next(Msa.FORBIDDEN)
 				await this.db.destroy({ where:{ key, num } })
 				// TODO rm votes
 				res.sendStatus(200)
